Code refactor, bug fix, created new DTO

+++ frontend/src/contexts/AuctionContext.tsx
@@ -11,9 +11,13 @@ import { State } from "../objects/Enums";
 
 import AuctionHttpRepository from "../repositories/implementations/AuctionHttpRepository";
 import { UserAuthorizationContext } from "./UserAuthorizationContext";
+import { useLocalStorage } from "@uidotdev/usehooks";
 
 interface IAuctionContext {
   auctions: Auction[] | undefined;
+
+  curAuctionId: string;
+  setAuctionId: (auctionId: string) => void;
 
   createAuction: (
     title: string,
@@ -22,6 +26,7 @@ interface IAuctionContext {
   ) => Promise<void>;
   getAuction: (id: string) => Promise<Auction | undefined>;
   deleteAuction: (auctionId: string) => Promise<void>;
+  changeState: (auctionId: string, state: State) => Promise<void>;
 }
 
 export const AuctionContext = createContext<IAuctionContext>(
@@ -34,10 +39,11 @@ export const AuctionProvider: React.FC<PropsWithChildren> = ({ children }) => {
   const [auctions, setAuctions] = useState<Auction[] | undefined>([]);
   const { user } = useContext(UserAuthorizationContext);
 
+  const [curAuctionId, saveAuctionId] = useLocalStorage("savedAuctionId", "");
+
   useEffect(() => {
     const fetchAuctions = async () => {
       if (!user) return;
-
       setAuctions(await auctionRepository.getAsync());
     };
 
@@ -63,19 +69,43 @@ export const AuctionProvider: React.FC<PropsWithChildren> = ({ children }) => {
     await auctionRepository.postAsync(auction);
   };
 
-  const getAuction = async (id: string): Promise<Auction | undefined> => {
-    const auction = await auctionRepository.getByIdAsync(id);
+  const getAuction = async (
+    auctionId: string
+  ): Promise<Auction | undefined> => {
+    const auction = await auctionRepository.getByIdAsync(auctionId);
     if (!auction) return;
     return auction;
   };
 
   const deleteAuction = async (auctionId: string) => {
+    if (!(await getAuction(auctionId))) return;
     await auctionRepository.deleteAsync(auctionId);
+  };
+
+  const changeState = async (auctionId: string, state: State) => {
+    if (!(await getAuction(auctionId))) return;
+
+    if (state == State.running)
+      await auctionRepository.setDateStartAsync(auctionId);
+
+    await auctionRepository.changeStateAsync(auctionId, state);
+  };
+
+  const setAuctionId = (auctionId: string) => {
+    saveAuctionId(auctionId);
   };
 
   return (
     <AuctionContext.Provider
-      value={{ auctions, createAuction, getAuction, deleteAuction }}
+      value={{
+        auctions,
+        curAuctionId,
+        setAuctionId,
+        createAuction,
+        getAuction,
+        deleteAuction,
+        changeState,
+      }}
     >
       {children}
     </AuctionContext.Provider>
